consider ungraded Assignment attempts when determining fail status
By somewhat popular request, when the Completion API returns 'completed-
failed' status (red), consider whether an ungraded reattempt exists and
pull it back to 'submitted' (yellow) if so. The displayed status in the
bar's details will be "Not completed (Submitted)", which is a lie, but
one that placates students. Note that the Activity Completion report is
still going to show the true "Completed (did not achieve pass grade)"
state of things in line with how the system actually works.

Fixes CONTRIB-8245: late submissions show red instead of yellow
Fixes #58: resubmissions show red instead of yellow