consider team-submit Assignments in their various configurations
This is somewhat of a 'perfect world' implementation of the logic
knowing that reality is anything but, and I'm not inclined to invoke the
heft of all Assignment's group/user logic with its resultant database
spam just for this. Students moving between groups after submissions are
made is already a sketchy situation, so students affected by such
anomalies will probably show skewed statuses in the bar.

Fixes #46: submission state when working with groupings
Fixes CONTRIB-8238: ... ignores group submissions in Assignment ...