revisit site-level page checking again
Assignment activity loads parts of its grader app using ajax via the
'core/fragment' module. This occurs with the pagetype of the 'page'
being rendered set to 'site-index', which is unusual and causes the
block to think it's being rendered on a dashboard. In a site with a
user enrolled in thousands of courses each with a block instance, a
lot of waste is caused by the block iterating over all the enrolled
courses just to produce a fragment of HTML.

Now, we just look at the page context to work out whether to produce
dashboard or course-level content.

And to be sure I haven't fouled up async course copy, that is now
tested.

Refs #83