don't give coursecreator role any capabilities by default
Given how the role's purpose is described in the documentation
(https://docs.moodle.org/405/en/Course_creator_role) it doesn't seem
wise to let them view the overview, or necessary to be able to add an
instance in keeping with how moodle/site:manageblocks is defined.
And to be safe, reset any of the formerly-allowed capabilities for the
role consistent with how they would be on a new install of the plugin.

Fixes CONTRIB-9800