improve the user experience for overview page sorting by percentage
We compute the percentages and store them in the DB for the table
renderer to query, providing a progress bar while doing all the maths.
Calculation is triggered if percentage is a sort key and used until
the cached value expires at the configured cache timeout or is early-
evicted if a student completes an activity. If calculation is never
triggered, percentages will computed real-time for the page of results
being displayed.

resolves #95